requirements.txt
asgiref==3.3.1
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.1.6
django-extensions==3.1.1
gunicorn==20.0.4
idna==3.2
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.26.0
simplejson==3.17.2
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
